Moved CatsAdapter to Adapters package
